# Direct check performance improvements