--- a/internal/datastore/memdb/readonly.go
+++ b/internal/datastore/memdb/readonly.go
@@ -5,6 +5,8 @@ import (
 	"fmt"
 	"runtime"
 
+	"github.com/authzed/spicedb/internal/datastore/common"
+
 	"github.com/hashicorp/go-memdb"
 	"github.com/jzelinskie/stringz"
 
@@ -50,7 +52,7 @@ func (r *memdbReader) QueryRelationships(
 		filter.ResourceType,
 		filter.OptionalResourceIds,
 		filter.OptionalResourceRelation,
-		filter.OptionalSubjectsFilter,
+		filter.OptionalSubjectsSelectors,
 		filter.OptionalCaveatName,
 		queryOpts.Usersets,
 	)
@@ -61,31 +63,8 @@ func (r *memdbReader) QueryRelationships(
 		limit: queryOpts.Limit,
 	}
 
-	runtime.SetFinalizer(iter, mustHaveBeenClosed)
+	runtime.SetFinalizer(iter, common.MustHaveBeenClosed)
 	return iter, nil
-}
-
-func mustHaveBeenClosed(iter *memdbTupleIterator) {
-	if !iter.closed {
-		panic("Tuple iterator garbage collected before Close() was called")
-	}
-}
-
-// QueryRelationshipsForDirectCheck reads relationships for a direct check operation.
-func (r *memdbReader) QueryRelationshipsForDirectCheck(
-	ctx context.Context,
-	filter datastore.DirectCheckRelationshipsFilter,
-	options ...options.QueryOptionsOption,
-) (datastore.RelationshipIterator, error) {
-	// NOTE: This does not do any form of optimized lookup for the direct check, but it is
-	// strictly always correct to return *more* of the relationships, so we do so here because
-	// memdb is not considered a datastore where highest performance is required.
-	// TODO(jschorr): Improve this if ever necessary.
-	return r.QueryRelationships(ctx, datastore.RelationshipsFilter{
-		ResourceType:             filter.ResourceType,
-		OptionalResourceIds:      filter.ResourceIds,
-		OptionalResourceRelation: filter.ResourceRelation,
-	})
 }
 
 // ReverseQueryRelationships reads relationships starting from the subject.
@@ -127,7 +106,7 @@ func (r *memdbReader) ReverseQueryRelationships(
 		filterObjectType,
 		nil,
 		filterRelation,
-		&subjectsFilter,
+		[]datastore.SubjectsSelector{subjectsFilter.AsSelector()},
 		"",
 		nil,
 	)
@@ -287,7 +266,7 @@ func filterFuncForFilters(
 	optionalResourceType string,
 	optionalResourceIds []string,
 	optionalRelation string,
-	optionalSubjectsFilter *datastore.SubjectsFilter,
+	optionalSubjectsSelectors []datastore.SubjectsSelector,
 	optionalCaveatFilter string,
 	usersets []*core.ObjectAndRelation,
 ) memdb.FilterFunc {
@@ -305,22 +284,40 @@ func filterFuncForFilters(
 			return true
 		}
 
-		if optionalSubjectsFilter != nil {
+		applySubjectSelector := func(selector datastore.SubjectsSelector) bool {
+			switch {
+			case len(selector.OptionalSubjectType) > 0 && selector.OptionalSubjectType != tuple.subjectNamespace:
+				return false
+			case len(selector.OptionalSubjectIds) > 0 && !stringz.SliceContains(selector.OptionalSubjectIds, tuple.subjectObjectID):
+				return false
+			}
+
+			if selector.RelationFilter.OnlyNonEllipsisRelations {
+				return tuple.subjectRelation != datastore.Ellipsis
+			}
+
 			relations := make([]string, 0, 2)
-			if optionalSubjectsFilter.RelationFilter.IncludeEllipsisRelation {
+			if selector.RelationFilter.IncludeEllipsisRelation {
 				relations = append(relations, datastore.Ellipsis)
 			}
 
-			if optionalSubjectsFilter.RelationFilter.NonEllipsisRelation != "" {
-				relations = append(relations, optionalSubjectsFilter.RelationFilter.NonEllipsisRelation)
+			if selector.RelationFilter.NonEllipsisRelation != "" {
+				relations = append(relations, selector.RelationFilter.NonEllipsisRelation)
 			}
 
-			switch {
-			case optionalSubjectsFilter.SubjectType != tuple.subjectNamespace:
-				return true
-			case len(optionalSubjectsFilter.OptionalSubjectIds) > 0 && !stringz.SliceContains(optionalSubjectsFilter.OptionalSubjectIds, tuple.subjectObjectID):
-				return true
-			case len(relations) > 0 && !stringz.SliceContains(relations, tuple.subjectRelation):
+			return len(relations) == 0 || stringz.SliceContains(relations, tuple.subjectRelation)
+		}
+
+		if len(optionalSubjectsSelectors) > 0 {
+			hasMatchingSelector := false
+			for _, selector := range optionalSubjectsSelectors {
+				if applySubjectSelector(selector) {
+					hasMatchingSelector = true
+					break
+				}
+			}
+
+			if !hasMatchingSelector {
 				return true
 			}
 		}