--- a/internal/datastore/mysql/reader.go
+++ b/internal/datastore/mysql/reader.go
@@ -47,6 +47,16 @@ var schema = common.SchemaInformation{
 	ColCaveatName:       colCaveatName,
 }
 
+func (mr *mysqlReader) QueryRelationshipsForDirectCheck(
+	ctx context.Context,
+	filter datastore.DirectCheckRelationshipsFilter,
+	opts ...options.QueryOptionsOption,
+) (iter datastore.RelationshipIterator, err error) {
+	// TODO (@vroldanbet) dupe from postgres datastore - need to refactor
+	qBuilder := common.NewSchemaQueryFilterer(schema, mr.filterer(mr.QueryTuplesQuery)).FilterWithDirectCheckFilter(filter)
+	return mr.querySplitter.SplitAndExecuteQuery(ctx, qBuilder, opts...)
+}
+
 func (mr *mysqlReader) QueryRelationships(
 	ctx context.Context,
 	filter datastore.RelationshipsFilter,