--- a/internal/datastore/proxy/observable.go
+++ b/internal/datastore/proxy/observable.go
@@ -211,8 +211,7 @@ func (i *observableRelationshipIterator) Close() {
 
 func (r *observableReader) ReverseQueryRelationships(ctx context.Context, subjectFilter datastore.SubjectsFilter, options ...options.ReverseQueryOptionsOption) (datastore.RelationshipIterator, error) {
 	ctx, closer := observe(ctx, "ReverseQueryRelationships")
-
-	iterator, err := r.delegate.ReverseQueryRelationships(ctx, subjectsFilter, options...)
+	iterator, err := r.delegate.ReverseQueryRelationships(ctx, subjectFilter, options...)
 	if err != nil {
 		return iterator, err
 	}