… .github/workflows/build-test.yaml
… e2e/go.sum
… internal/datastore/common/sql_test.go
… internal/datastore/proxy/proxy_test/mock.go
… internal/dispatch/graph/check_test.go
… internal/graph/check.go
… internal/graph/membershipset.go
… internal/graph/reachableresources.go
… internal/services/integrationtesting/benchconfigs/checkwidedirect.yaml
… internal/services/integrationtesting/benchmark_test.go
… internal/services/integrationtesting/consistency_datastore_test.go
… internal/services/shared/schema.go
… internal/testfixtures/validating.go
… pkg/datastore/datastore.go
… pkg/datastore/datastore_test.go
… pkg/datastore/relationshipquerytree.go
… pkg/datastore/test/datastore.go
… pkg/datastore/test/tuples.go
… pkg/validationfile/loader.go