Drastically improve the performance of "direct" checks on relations (e.g. when looking for a terminal subject in a relation with many such terminal subjects) by creating more specific and targeted queries

Performance comparison:

```
name                                                     old time/op    new time/op    
Services/basic_check_for_a_user/cockroachdb/v1-10          56.9ms         56.0ms        
Services/basic_check_for_a_user/mysql/v1-10                68.4ms         74.0ms        
Services/basic_check_for_a_user/postgres/v1-10             29.8ms         33.5ms        
Services/basic_check_for_a_user/spanner/v1-10              29.7ms         22.2ms        
Services/recursive_check_for_a_user/cockroachdb/v1-10       137ms          177ms        
Services/recursive_check_for_a_user/mysql/v1-10             191ms          179ms        
Services/recursive_check_for_a_user/postgres/v1-10         96.9ms        111.4ms        
Services/recursive_check_for_a_user/spanner/v1-10          62.6ms         63.4ms        
Services/wide_groups_check_for_a_user/cockroachdb/v1-10     108ms          126ms        
Services/wide_groups_check_for_a_user/mysql/v1-10           129ms          152ms        
Services/wide_groups_check_for_a_user/postgres/v1-10       75.3ms        100.6ms        
Services/wide_groups_check_for_a_user/spanner/v1-10        55.3ms         71.5ms        

Services/wide_direct_relation_check/cockroachdb/v1-10      73.1ms         33.6ms        
Services/wide_direct_relation_check/mysql/v1-10             200ms           36ms        
Services/wide_direct_relation_check/postgres/v1-10         62.8ms         21.4ms        
Services/wide_direct_relation_check/spanner/v1-10           137ms           47ms        

name                                                     old alloc/op   new alloc/op   
Services/basic_check_for_a_user/cockroachdb/v1-10          1.12MB         1.55MB        
Services/basic_check_for_a_user/mysql/v1-10                1.00MB         1.11MB        
Services/basic_check_for_a_user/postgres/v1-10             1.25MB         1.79MB        
Services/basic_check_for_a_user/spanner/v1-10              2.06MB         1.72MB        
Services/recursive_check_for_a_user/cockroachdb/v1-10      1.69MB         1.70MB        
Services/recursive_check_for_a_user/mysql/v1-10             825kB         2217kB        
Services/recursive_check_for_a_user/postgres/v1-10         3.39MB         1.93MB        
Services/recursive_check_for_a_user/spanner/v1-10          2.72MB         3.16MB        
Services/wide_groups_check_for_a_user/cockroachdb/v1-10    2.46MB         1.63MB        
Services/wide_groups_check_for_a_user/mysql/v1-10          1.49MB         1.40MB        
Services/wide_groups_check_for_a_user/postgres/v1-10       1.36MB         1.57MB        
Services/wide_groups_check_for_a_user/spanner/v1-10        1.74MB         2.06MB        

Services/wide_direct_relation_check/cockroachdb/v1-10      9.40MB         1.09MB        
Services/wide_direct_relation_check/mysql/v1-10            8.56MB         1.20MB        
Services/wide_direct_relation_check/postgres/v1-10         8.98MB         1.11MB        
Services/wide_direct_relation_check/spanner/v1-10          18.1MB          1.4MB        

name                                                     old allocs/op  new allocs/op  
Services/basic_check_for_a_user/cockroachdb/v1-10           4.18k          4.30k        
Services/basic_check_for_a_user/mysql/v1-10                 4.63k          4.73k        
Services/basic_check_for_a_user/postgres/v1-10              4.71k          4.84k        
Services/basic_check_for_a_user/spanner/v1-10               7.35k          7.19k        
Services/recursive_check_for_a_user/cockroachdb/v1-10       13.8k          14.1k        
Services/recursive_check_for_a_user/mysql/v1-10             15.5k          15.7k        
Services/recursive_check_for_a_user/postgres/v1-10          15.7k          16.0k        
Services/recursive_check_for_a_user/spanner/v1-10           23.0k          23.3k        
Services/wide_groups_check_for_a_user/cockroachdb/v1-10     10.0k          11.2k        
Services/wide_groups_check_for_a_user/mysql/v1-10           11.7k          12.9k        
Services/wide_groups_check_for_a_user/postgres/v1-10        10.7k          12.1k        
Services/wide_groups_check_for_a_user/spanner/v1-10         18.6k          21.6k        

Services/wide_direct_relation_check/cockroachdb/v1-10        119k             2k        
Services/wide_direct_relation_check/mysql/v1-10              191k             2k        
Services/wide_direct_relation_check/postgres/v1-10           119k             2k        
Services/wide_direct_relation_check/spanner/v1-10            274k             4k        
```

> **Note**
> The benchmarks were generated by running the various datastores in containers on a local M1, which likely explains why Postgres is so much faster: it is running under native `arm`. MySQL, CRDB, and Spanner are `amd64` images and therefore likely much slower than a real-world installation

Fixes #752